refactor: moving credit card class to the models folder
